[C++ builds] Factor ServerSessionMiddleware to separate header for testability
