+++ cpp/src/arrow/flight/sql/server_session_middleware.h
@@ -21,6 +21,7 @@
 #pragma once
 
 #include <functional>
+#include <map>
 #include <optional>
 #include <shared_mutex>
 #include <string_view>
